// server/types/notification/enum/SocketEventEnum.ts
enum SocketEventEnum {
  USER_CONNECTION = 'USER_CONNECTION',
  CREW_CONNECTION = 'CREW_CONNECTION',
  MAKE_SCHEDULE = 'MAKE_SCHEDULE',
  SEND_NOTIFICATION = 'SEND_NOTIFICATION',
  ACTIVE_CAR_UPDATE = 'ACTIVE_CAR_UPDATE',
  CAR_UPDATE = 'CAR_UPDATE',
  CAR_REMOVE = 'CAR_REMOVE',
  CAR_ADD = 'CAR_ADD',
  SWIP_SHIFT = 'SWIP_SHIFT',
  SEND_MESSAGE_TO_CHAT = 'SEND_MESSAGE_TO_CHAT',
  CHANGE_IS_ROUND_FUEL = 'CHANGE_IS_ROUND_FUEL',
  CHANGE_FUEL_SEASON = 'CHANGE_FUEL_SEASON',
  USER_PROFILE_UPDATE = 'USER_PROFILE_UPDATE',
  DISCONNECT = 'DISCONNECT',
}

export default SocketEventEnum;